refactor: 예외 처리 기능 수정
