+++ app/src/main/java/woowacourse/shopping/data/datasource/impl/ProductRemoteDataSourceImpl.kt
@@ -3,9 +3,8 @@ package woowacourse.shopping.data.datasource.impl
 import woowacourse.shopping.data.datasource.ProductRemoteDataSource
 import woowacourse.shopping.data.dto.response.ResponseProductIdGetDto
 import woowacourse.shopping.data.dto.response.ResponseProductsGetDto
+import woowacourse.shopping.data.exception.ShoppingError
 import woowacourse.shopping.data.remote.service.ProductService
-import woowacourse.shopping.exception.ShoppingError
-import woowacourse.shopping.exception.ShoppingException
 
 class ProductRemoteDataSourceImpl(private val service: ProductService) : ProductRemoteDataSource {
     override suspend fun getProductsByOffset(
@@ -13,9 +12,8 @@ class ProductRemoteDataSourceImpl(private val service: ProductService) : Product
         size: Int,
     ): Result<ResponseProductsGetDto> =
         runCatching {
-            service.getProductsByOffset(page = page, size = size).body() ?: throw ShoppingException(
-                ShoppingError.ProductNotFound,
-            )
+            service.getProductsByOffset(page = page, size = size).body()
+                ?: throw ShoppingError.ProductNotFound
         }
 
     override suspend fun getProductsByCategory(
@@ -24,12 +22,12 @@ class ProductRemoteDataSourceImpl(private val service: ProductService) : Product
     ): Result<ResponseProductsGetDto> =
         runCatching {
             service.getProductsByCategory(category = category, page = page).body()
-                ?: throw ShoppingException(ShoppingError.ProductNotFound)
+                ?: throw ShoppingError.ProductNotFound
         }
 
     override suspend fun getProductsById(id: Long): Result<ResponseProductIdGetDto> =
         runCatching {
             service.getProductsById(id = id).body()
-                ?: throw ShoppingException(ShoppingError.ProductNotFoundWithId(id))
+                ?: throw ShoppingError.ProductNotFoundWithId(id)
         }
 }
