--- app/src/main/java/woowacourse/shopping/exception/ShoppingException.kt
@@ -1,3 +1,0 @@
-package woowacourse.shopping.exception
-
-class ShoppingException(val error: ShoppingError) : Exception()
