+++ app/src/main/java/woowacourse/shopping/ui/products/ProductContentsActivity.kt
@@ -18,7 +18,6 @@ import woowacourse.shopping.data.repository.CartRepositoryImpl
 import woowacourse.shopping.data.repository.ProductRepositoryImpl
 import woowacourse.shopping.data.repository.RecentProductRepositoryImpl
 import woowacourse.shopping.databinding.ActivityProductContentsBinding
-import woowacourse.shopping.exception.handleError
 import woowacourse.shopping.ui.cart.CartActivity
 import woowacourse.shopping.ui.detail.ProductDetailActivity
 import woowacourse.shopping.ui.products.adapter.ProductAdapter
@@ -66,8 +65,7 @@ class ProductContentsActivity : AppCompatActivity() {
 
     private fun observeError() {
         viewModel.error.observe(this) {
-            val errMsg = handleError(it)
-            toast = Toast.makeText(this, errMsg, Toast.LENGTH_SHORT)
+            toast = Toast.makeText(this, it.message, Toast.LENGTH_SHORT)
             toast?.show()
         }
     }
